Install OpenEBS
In order to run Percona, we need to configure Persistent Volumes. While we can configure replicated storage (like longhorn or Ceph), local storage will do fine since the MySQL operator will ensure durability through replication. In the event that a node completely fails, the operator will spin up a new MySQL instance on a new node and using EBS will dynamically provision a PV and a PVC and bring that instance up to speed with the other instances.