Adds Cilium's Gateway API
We need Gateway API to expose services outside of the cluster. While we could go ahead with other implementations like Istio, we chose Cilium since it's already a dependency of the project which reduces the operational overhead of security patches.
This diff enables cilium's Gateway API.

NOTE: This diff isn't complete on its own without the resolution of [this](https://github.com/issues/created?issue=hcloud-talos%7Cterraform-hcloud-talos%7C278) issue. If it's not resolved, then we need to create a cilium default values yaml.