Add grafana secret to myql namespace
Secret containing the mysql grafana user's password was moved to the namespace monitoring after the user was created in MySQL. After destroying and rebuilding the cluster, mysql didn't pick up the secret since it was in a different namespace.

While we could figure out a synchronisation logic, since this secret wouldn't change much, this diff copies the secret to the mysql namespace.