Provisions Loki
This diff adds Loki which is a log storage that uses the existing Minio cluster.

We created minio user and policy using the minio runbook. While ideally, Loki should run in a different tier from production, in the first iteration, we can keep it on Minio's prod tier for simplicity.