Disable Loki Persistence
Loki failed to schedule because it has an unmet PVC. While we could provision a PVC for it, it doesn't seem imminently necessary according to [this](https://github.com/grafana/loki/issues/7068) issue. It seems that storage is only required to park logs in case minio is down. Since provisioning such storage can't simply use EBS, we will disable this feature for now and revisit when needed.