[muzbuild] Change Mixpanel secret
This diff changes mixpanel secret to point to the prod project. This was done as part of discovering that the endpoint exposing both stream token and mixpanel key didn't require authentication.

We only replace mixpanel here as stream token is pretty much useless without stream secret or a valid token