[caas_dev] Add Caas Dev job
In order to properly test infrastructure-heavy components like video transcoding or CDN delivery, we need to test it in a kubernetes environment mimicking the production environment. Ideally, we should have an entire dev cluster. However, since it's too costly right now, we just deploy a dev job for caas backend which we will connect a local frontend to or use MTLS with a dev frontend.