[minio] Add Caas dev user
This diff adds a dev user for caas dev. We are using the same deployment as ACLs allow us to have control separation. The only side effect is that dev can affect production traffic. Since we don't plan to do any performance extensive tests, this setup should suffice for now.